inject/requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
idna==2.8
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
parso==0.5.1
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.1
prompt-toolkit==3.0.2
Pygments==2.5.2
pymongo==3.9.0
pyparsing==2.4.5
PyPDF2==1.26.0
pytesseract==0.3.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
six==1.13.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wincertstore==0.2
